CMake 3.6.3

# gpg: Signature made Thu Nov  3 09:39:28 2016 CDT
# gpg:                using RSA key 2D2CEF1034921684
# gpg:                issuer "brad.king@kitware.com"
# gpg: Can't check signature: No public key

+++ Source/kwsys/SystemTools.cxx
@@ -35,10 +35,12 @@
 #include <fstream>
 #include <sstream>
 #include <set>
+#include <vector>
 
 // Work-around CMake dependency scanning limitation.  This must
 // duplicate the above list of headers.
 #if 0
+# include "RegularExpression.hxx.in"
 # include "SystemTools.hxx.in"
 # include "Directory.hxx.in"
 # include "FStream.hxx.in"
@@ -87,6 +89,7 @@
 // Windows API.
 #if defined(_WIN32)
 # include <windows.h>
+# include <winioctl.h>
 # ifndef INVALID_FILE_ATTRIBUTES
 #  define INVALID_FILE_ATTRIBUTES ((DWORD)-1)
 # endif
@@ -2400,93 +2403,93 @@ bool SystemTools::CopyFileAlways(const std::string& source, const std::string& d
     }
   else
     {
-    const int bufferSize = 4096;
-    char buffer[bufferSize];
+  const int bufferSize = 4096;
+  char buffer[bufferSize];
 
-    // If destination is a directory, try to create a file with the same
-    // name as the source in that directory.
+  // If destination is a directory, try to create a file with the same
+  // name as the source in that directory.
 
     std::string destination_dir;
     if(SystemTools::FileIsDirectory(destination))
-      {
-      destination_dir = real_destination;
-      SystemTools::ConvertToUnixSlashes(real_destination);
-      real_destination += '/';
+    {
+    destination_dir = real_destination;
+    SystemTools::ConvertToUnixSlashes(real_destination);
+    real_destination += '/';
       std::string source_name = source;
-      real_destination += SystemTools::GetFilenameName(source_name);
+    real_destination += SystemTools::GetFilenameName(source_name);
+    }
+  else
+    {
+    destination_dir = SystemTools::GetFilenamePath(destination);
+    }
+
+  // Create destination directory
+
+  SystemTools::MakeDirectory(destination_dir);
+
+  // Open files
+#if defined(_WIN32)
+  kwsys::ifstream fin(Encoding::ToNarrow(
+    SystemTools::ConvertToWindowsExtendedPath(source)).c_str(),
+                  std::ios::in | std::ios::binary);
+#else
+  kwsys::ifstream fin(source.c_str(),
+                  std::ios::in | std::ios::binary);
+#endif
+  if(!fin)
+    {
+    return false;
+    }
+
+  // try and remove the destination file so that read only destination files
+  // can be written to.
+  // If the remove fails continue so that files in read only directories
+  // that do not allow file removal can be modified.
+  SystemTools::RemoveFile(real_destination);
+
+#if defined(_WIN32)
+  kwsys::ofstream fout(Encoding::ToNarrow(
+    SystemTools::ConvertToWindowsExtendedPath(real_destination)).c_str(),
+                     std::ios::out | std::ios::trunc | std::ios::binary);
+#else
+  kwsys::ofstream fout(real_destination.c_str(),
+                     std::ios::out | std::ios::trunc | std::ios::binary);
+#endif
+  if(!fout)
+    {
+    return false;
+    }
+
+  // This copy loop is very sensitive on certain platforms with
+  // slightly broken stream libraries (like HPUX).  Normally, it is
+  // incorrect to not check the error condition on the fin.read()
+  // before using the data, but the fin.gcount() will be zero if an
+  // error occurred.  Therefore, the loop should be safe everywhere.
+  while(fin)
+    {
+    fin.read(buffer, bufferSize);
+    if(fin.gcount())
+      {
+      fout.write(buffer, fin.gcount());
       }
     else
       {
-      destination_dir = SystemTools::GetFilenamePath(destination);
+      break;
       }
+    }
 
-    // Create destination directory
+  // Make sure the operating system has finished writing the file
+  // before closing it.  This will ensure the file is finished before
+  // the check below.
+  fout.flush();
 
-    SystemTools::MakeDirectory(destination_dir);
+  fin.close();
+  fout.close();
 
-    // Open files
-#if defined(_WIN32)
-    kwsys::ifstream fin(Encoding::ToNarrow(
-      SystemTools::ConvertToWindowsExtendedPath(source)).c_str(),
-                  std::ios::in | std::ios::binary);
-#else
-    kwsys::ifstream fin(source.c_str(),
-                  std::ios::in | std::ios::binary);
-#endif
-    if(!fin)
-      {
-      return false;
-      }
-
-    // try and remove the destination file so that read only destination files
-    // can be written to.
-    // If the remove fails continue so that files in read only directories
-    // that do not allow file removal can be modified.
-    SystemTools::RemoveFile(real_destination);
-
-#if defined(_WIN32)
-    kwsys::ofstream fout(Encoding::ToNarrow(
-      SystemTools::ConvertToWindowsExtendedPath(real_destination)).c_str(),
-                     std::ios::out | std::ios::trunc | std::ios::binary);
-#else
-    kwsys::ofstream fout(real_destination.c_str(),
-                     std::ios::out | std::ios::trunc | std::ios::binary);
-#endif
-    if(!fout)
-      {
-      return false;
-      }
-
-    // This copy loop is very sensitive on certain platforms with
-    // slightly broken stream libraries (like HPUX).  Normally, it is
-    // incorrect to not check the error condition on the fin.read()
-    // before using the data, but the fin.gcount() will be zero if an
-    // error occurred.  Therefore, the loop should be safe everywhere.
-    while(fin)
-      {
-      fin.read(buffer, bufferSize);
-      if(fin.gcount())
-        {
-        fout.write(buffer, fin.gcount());
-        }
-      else
-        {
-        break;
-        }
-      }
-
-    // Make sure the operating system has finished writing the file
-    // before closing it.  This will ensure the file is finished before
-    // the check below.
-    fout.flush();
-
-    fin.close();
-    fout.close();
-
-    if(!fout)
-      {
-      return false;
-      }
+  if(!fout)
+    {
+    return false;
+    }
     }
   if ( perms )
     {
@@ -2759,6 +2762,107 @@ std::string SystemTools::GetLastSystemError()
   return strerror(e);
 }
 
+#ifdef _WIN32
+
+static bool IsJunction(const std::wstring& source)
+{
+#ifdef FSCTL_GET_REPARSE_POINT
+  const DWORD JUNCTION_ATTRS = FILE_ATTRIBUTE_DIRECTORY |
+                               FILE_ATTRIBUTE_REPARSE_POINT;
+  DWORD attrs = GetFileAttributesW(source.c_str());
+  if (attrs == INVALID_FILE_ATTRIBUTES)
+    {
+    return false;
+    }
+  if ((attrs & JUNCTION_ATTRS) != JUNCTION_ATTRS)
+    {
+    return false;
+    }
+
+  // Adjust privileges so that we can succefully open junction points.
+  HANDLE token;
+  TOKEN_PRIVILEGES privs;
+  OpenProcessToken(GetCurrentProcess(), TOKEN_ADJUST_PRIVILEGES, &token);
+  LookupPrivilegeValue(NULL, SE_BACKUP_NAME, &privs.Privileges[0].Luid);
+  privs.PrivilegeCount = 1;
+  privs.Privileges[0].Attributes = SE_PRIVILEGE_ENABLED;
+  AdjustTokenPrivileges(token, FALSE, &privs, sizeof(TOKEN_PRIVILEGES), NULL, NULL);
+  CloseHandle(token);
+
+  HANDLE dir = CreateFileW(source.c_str(), GENERIC_READ,
+                           0, NULL, OPEN_EXISTING,
+                           FILE_FLAG_OPEN_REPARSE_POINT | FILE_FLAG_BACKUP_SEMANTICS, NULL);
+  if (dir == INVALID_HANDLE_VALUE)
+    {
+    return false;
+    }
+
+  // Query whether this is a reparse point or not.
+  BYTE buffer[MAXIMUM_REPARSE_DATA_BUFFER_SIZE];
+  REPARSE_GUID_DATA_BUFFER *reparse_buffer =
+    (REPARSE_GUID_DATA_BUFFER*) buffer;
+  DWORD sentinel;
+
+  BOOL success = DeviceIoControl(
+    dir, FSCTL_GET_REPARSE_POINT,
+    NULL, 0,
+    reparse_buffer, MAXIMUM_REPARSE_DATA_BUFFER_SIZE,
+    &sentinel, NULL);
+
+  CloseHandle(dir);
+
+  return (success && (reparse_buffer->ReparseTag == IO_REPARSE_TAG_MOUNT_POINT));
+#else
+  return false;
+#endif
+}
+
+static bool DeleteJunction(const std::wstring& source)
+{
+#ifdef FSCTL_DELETE_REPARSE_POINT
+  // Adjust privileges so that we can succefully open junction points as
+  // read/write.
+  HANDLE token;
+  TOKEN_PRIVILEGES privs;
+  OpenProcessToken(GetCurrentProcess(), TOKEN_ADJUST_PRIVILEGES, &token);
+  LookupPrivilegeValue(NULL, SE_RESTORE_NAME, &privs.Privileges[0].Luid);
+  privs.PrivilegeCount = 1;
+  privs.Privileges[0].Attributes = SE_PRIVILEGE_ENABLED;
+  AdjustTokenPrivileges(token, FALSE, &privs, sizeof(TOKEN_PRIVILEGES), NULL, NULL);
+  CloseHandle(token);
+
+  HANDLE dir = CreateFileW(source.c_str(), GENERIC_READ | GENERIC_WRITE,
+                           0, NULL, OPEN_EXISTING,
+                           FILE_FLAG_OPEN_REPARSE_POINT | FILE_FLAG_BACKUP_SEMANTICS, NULL);
+  if (dir == INVALID_HANDLE_VALUE)
+    {
+    return false;
+    }
+
+  // Set up the structure so that we can delete the junction.
+  std::vector<BYTE> buffer(REPARSE_GUID_DATA_BUFFER_HEADER_SIZE, 0);
+  REPARSE_GUID_DATA_BUFFER *reparse_buffer =
+    (REPARSE_GUID_DATA_BUFFER*) &buffer[0];
+  DWORD sentinel;
+
+  reparse_buffer->ReparseTag = IO_REPARSE_TAG_MOUNT_POINT;
+
+  BOOL success = DeviceIoControl(
+    dir, FSCTL_DELETE_REPARSE_POINT,
+    reparse_buffer, REPARSE_GUID_DATA_BUFFER_HEADER_SIZE,
+    NULL, 0,
+    &sentinel, NULL);
+
+  CloseHandle(dir);
+
+  return !!success;
+#else
+  return false;
+#endif
+}
+
+#endif
+
 bool SystemTools::RemoveFile(const std::string& source)
 {
 #ifdef _WIN32
@@ -2785,6 +2889,10 @@ bool SystemTools::RemoveFile(const std::string& source)
     {
     SetLastError(err);
     return false;
+    }
+  if (IsJunction(ws) && DeleteJunction(ws))
+    {
+    return true;
     }
   if (DeleteFileW(ws.c_str()) ||
       GetLastError() == ERROR_FILE_NOT_FOUND ||
@@ -2985,17 +3093,17 @@ std::string SystemTools::FindProgram(
     extensions.push_back(".com");
     extensions.push_back(".exe");
 
-    // first try with extensions if the os supports them
+  // first try with extensions if the os supports them
     for(std::vector<std::string>::iterator i =
-          extensions.begin(); i != extensions.end(); ++i)
-      {
-      tryPath = name;
-      tryPath += *i;
+        extensions.begin(); i != extensions.end(); ++i)
+    {
+    tryPath = name;
+    tryPath += *i;
       if(SystemTools::FileExists(tryPath, true))
-        {
-        return SystemTools::CollapseFullPath(tryPath);
-        }
+      {
+      return SystemTools::CollapseFullPath(tryPath);
       }
+    }
     }
 #endif
 
@@ -4656,8 +4764,9 @@ bool SystemTools::GetLineFromStream(std::istream& is,
   // been reached.  Clear the fail bit just before reading.
   while(!haveNewline &&
         leftToRead != 0 &&
-        (is.clear(is.rdstate() & ~std::ios::failbit),
-         is.getline(buffer, bufferSize), is.gcount() > 0))
+        (static_cast<void>(is.clear(is.rdstate() & ~std::ios::failbit)),
+         static_cast<void>(is.getline(buffer, bufferSize)),
+         is.gcount() > 0))
     {
     // We have read at least one byte.
     haveData = true;
